Update buttonmap.xml and topology.xml
This is a best-effort attempt to improve BlueMSX's buttonmap and
topology. BlueMSX doesn't conform to libretro's set_controller API, so
the best we can do for now is default to a ColecoVision controller for
all games, and rely on BlueMSX to disregard this and poll for MSX joystick
input for non-Coleco games.